BUG: Rename _rpca to _standalone_rpca
fixes a bug in the refactoring done in the prior commit. Now,
all the tests pass.

Potentially worth noting that the test output (at least for the
standalone RPCA) looks slightly different from what's in the
repository. However, it looks like the values are only slightly
different (it's probably due to numpy version diffs/machine precision
diffs or something), and all the tests pass now, so I think this
should be ok (but this isn't my call to make :)).

At this point, I think this branch might be ready to merge back
into biocore. But I'd like to see what y'all have to say about it.