TST: Make standalone RPCA test check correct files
It's failing, due to the "flipping" bug (?) and other small
differences between values as discussed in #29. The "truth" files
might just be really old.

I also added in "truth_sample_trimmed.txt", which is just a version
of the truth_sample.txt file without the OrdinationResults-ish stuff
at the end (so you can read it in pandas easier).